Move format_datetime to base.utils to break import cycle
